+++ docs/changes.rst
@@ -6,6 +6,9 @@ Changelog
 
 - Statistics: Fix updating tool statistics.
   [reinhardt]
+- Statistics: Add newsletter statistics.
+  (`#925 <https://github.com/syslabcom/scrum/issues/925>`_).
+  [reinhardt]
 - Statistics: Exclude accounts with certain domains
   (`#923 <https://github.com/syslabcom/scrum/issues/923>`_).
   [reinhardt]